API: Move submit button in form
